requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.2.1=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2022.6.15=h5b45459_0
certifi=2022.6.15=pyhd8ed1ab_1
cffi=1.15.1=py38h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=py38haa95532_0
cryptography=37.0.1=py38h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py38h294d835_0
dask-core=2022.9.0=pyhd8ed1ab_0
enum34=1.1.10=py38h32f6830_2
fonttools=4.25.0=pyhd3eb1b0_0
freeglut=3.2.2=h0e60522_1
freetype=2.10.4=hd328e21_0
fsspec=2022.8.2=pyhd8ed1ab_0
gettext=0.19.8.1=ha2e2712_1008
glib=2.72.1=h7755175_0
glib-tools=2.72.1=h7755175_0
gst-plugins-base=1.20.3=h001b923_1
gstreamer=1.20.3=h6b5321d_1
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=70.1=h0e60522_0
idna=3.3=pyhd3eb1b0_0
imagecodecs-lite=2019.12.3=py38hbdcd294_5
imageio=2.21.2=pyhfa7a67d_0
intel-openmp=2021.4.0=haa95532_3556
jasper=2.0.33=h77af90b_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.2=py38hd77b12b_0
krb5=1.19.3=h1176d77_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=14.0.6=default_h77d9078_0
libclang13=14.0.6=default_h77d9078_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libglib=2.72.1=h3be07f2_0
libiconv=1.16=h2bbff1b_2
liblapack=3.9.0=5_hd5c7e75_netlib
liblapacke=3.9.0=5_hd5c7e75_netlib
libogg=1.3.5=h2bbff1b_1
libopencv=4.6.0=py38h3d2d820_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h7755175_1
libtiff=4.4.0=h8a3f274_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.2=h2bbff1b_0
libwebp-base=1.2.2=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzlib=1.2.12=h8ffe710_2
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.5.2=py38haa95532_0
matplotlib-base=3.5.2=py38hd77b12b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
networkx=2.8.6=pyhd8ed1ab_0
numexpr=2.8.3=py38hb80d3ca_0
numpy=1.23.1=py38h7a0a035_0
numpy-base=1.23.1=py38hca35cd5_0
opencv=4.6.0=py38haa244fe_0
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py38hd77b12b_0
partd=1.3.0=pyhd8ed1ab_0
pathlib=1.0.1=py38haa244fe_6
pcre=8.45=hd77b12b_0
pillow=9.2.0=py38hdc2b20a_1
pip=22.1.2=py38haa95532_0
ply=3.11=py38_0
py-opencv=4.6.0=py38h4317176_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38haa95532_0
pywavelets=1.3.0=py38hbdcd294_1
pyyaml=6.0=py38h294d835_4
qt-main=5.15.4=h467ea89_2
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h0db62b3_6
qudida=0.0.4=pyhd8ed1ab_0
requests=2.28.1=py38haa95532_0
scikit-image=0.19.3=py38hcc40339_1
scikit-learn=1.1.2=py38hc27f28a_0
scipy=1.9.1=py38h91810f7_0
setuptools=63.4.1=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h2bbff1b_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2019.7.26.2=py38_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=0.12.1=py38_cu113
torchvision=0.13.1=py38_cu113
tornado=6.1=py38h2bbff1b_0
tqdm=4.64.0=py38haa95532_0
typing-extensions=4.3.0=py38haa95532_0
typing_extensions=4.3.0=py38haa95532_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.11=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.29.30139=h890b9b1_7
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=h8ffe710_2
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
